+++ R/import.R
@@ -3,9 +3,12 @@
 #'                             parse_chartcode
 #' @importFrom chartplotter    process_chart
 #' @importFrom curl            curl
+#' @importFrom growthscreener  screen_curves_ind
+#' @importFrom jamesclient     upload_bds get_url
+#' @importFrom jsonlite        toJSON unbox
 #' @importFrom methods         new slot
-#' @importFrom minihealth      get_range convert_bds_individual
-#'                             is.individual
+#' @importFrom minihealth      convert_bds_individual
+#'                             get_range
 #' @importClassesFrom minihealth individual xyz bse
 NULL